requirements.txt
Jinja2>=3.1.0
Flask>=3.0.0
PyYAML>=6.0
